@@ -16,8 +16,8 @@
         }
     ],
     "require": {
-        "php": "7.4.*|^8.0",
-        "illuminate/contracts": "^8.37",
+        "php": "^8.1",
+        "laravel/framework": "^8.0|^9.0|^10.0",
         "intervention/image": "^2.7",
         "spatie/laravel-package-tools": "^1.4.3"
     },
@@ -48,7 +48,10 @@
         "test-coverage": "vendor/bin/phpunit --coverage-html coverage"
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
+        }
     },
     "extra": {
         "laravel": {
