@@ -2,223 +2,162 @@
 
 use EyadHamza\LaravelWebp\Exceptions\NoImageGivenException;
 use EyadHamza\LaravelWebp\Exceptions\NotImageException;
-use EyadHamza\LaravelWebp\ImageToWebp;
 
+use EyadHamza\LaravelWebp\Services\WebpService;
+use EyadHamza\LaravelWebp\Support\PathConversionSupport;
 use EyadHamza\LaravelWebp\Tests\TestSupport\Models\TestModel;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Artisan;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use Intervention\Image\ImageManagerStatic as Image;
+
 use function Pest\Laravel\assertDatabaseHas;
-use function Pest\Laravel\assertDatabaseMissing;
-use function Pest\Laravel\withoutExceptionHandling;
 
 beforeEach(function () {
-    TestModel::factory()->create([
-        'image' => asset('/storage/'.'test.jpg'),
-        'avatar' => asset('/storage/'.'test2.jpg'),
-        ]);
-    $this->refreshAndClean();
-});
+    Storage::fake();
 
-it('prepare test', function () {
-    expect(true)->toBeTrue();
-});
-
-it('can save an image', function () {
-    $testImage = TestModel::find(1);
-
-    $testImage->saveImageAsWebp();
-
-    Storage::disk()
-        ->assertExists($this->getTestImageWebpRelativePath());
-
-    Storage::disk()
-        ->assertExists(ImageToWebp::getOldImageRelativePath());
-});
-it('can save an image with a very small size', function () {
-    $testImage = TestModel::find(1);
-    $testImage->image = asset('/storage/'.'test3.png');
+    $this->image = UploadedFile::fake()
+        ->image('image.jpg')
+        ->store('public');
 
 
-    $testImage->saveImageAsWebp();
+    $this->webpImage = Str::replace('.jpg', '.webp', $this->image);
 
-    Storage::disk()
-        ->assertExists('public/test3.webp');
-
-    Storage::disk()
-        ->assertExists(ImageToWebp::getOldImageRelativePath());
-});
-
-it('can save an image with the right name', function () {
-    $testImage = TestModel::find(1);
-
-    $testImage->saveImageAsWebp();
-
-    $testImage->saveImageAsWebp();
-    Storage::disk()
-        ->assertExists($this->getTestImageRelativePath());
-
-    Storage::disk()
-        ->assertExists($this->getTestImageWebpRelativePath());
-
-    assertDatabaseHas('test_images', [
-        'image' => ImageToWebp::toFullPath($this->getTestImageWebpRelativePath()),
+    $this->testModel = TestModel::factory()->create([
+        'image' => $this->image,
     ]);
 });
 
-it('can overwrite an image', function () {
-    withoutExceptionHandling();
-    $testImage = TestModel::find(1);
+it('can cast an image to webp and store it on disk', function () {
+    Storage::disk()->assertExists($this->testModel->image);
 
-    $testImage->overwriteImageAsWebp();
-
-    Storage::disk()
-        ->assertExists($this->getTestImageWebpRelativePath());
-
-    Storage::disk()
-        ->assertMissing(ImageToWebp::getOldImageRelativePath());
+    Str::endsWith($this->testModel->image, '.webp');
 });
 
-it('must modify image url in the database', function () {
-    $testImage = TestModel::find(1);
-
-    $testImage->saveImageAsWebp();
-
-    assertDatabaseHas('test_images', [
-       'image' => ImageToWebp::toFullPath($this->getTestImageWebpRelativePath()),
-   ]);
-});
-
-it('can  save an image by passing the path', function () {
-    $testImage = TestModel::find(1);
-
-    ImageToWebp::make($testImage->image)->save();
+it('can overwrite the old image as the default behavior', function () {
+    Storage::disk()
+        ->assertExists($this->testModel->image);
 
     Storage::disk()
-        ->assertExists(ImageToWebp::getOldImageRelativePath());
+        ->assertMissing($this->image);
 });
 
-it('can overwrite an image by passing the path', function () {
-    $testImage = TestModel::find(1);
+it('can prevent overwrite the old image as the user config behavior', function () {
+    config()->set('webp.overwrite', false);
 
-    ImageToWebp::make($testImage->image)->overwrite();
+    $this->secondImage = UploadedFile::fake()
+        ->image('image.jpg')
+        ->store('public');
 
-    Storage::disk()
-        ->assertExists(ImageToWebp::getWebpRelativePath(ImageToWebp::toRelativePath($testImage->image)));
+    $this->secondWebpImage = Str::replace('.jpg', '.webp', $this->secondImage);
 
-    Storage::disk()
-        ->assertMissing(ImageToWebp::getOldImageRelativePath());
-});
-
-it('must modify image url in the database by passing the path', function () {
-    $testImage = TestModel::find(1);
-
-    $testImage->saveImageAsWebp($testImage->image);
-
-    assertDatabaseHas('test_images', [
-        'image' => ImageToWebp::toFullPath($this->getTestImageWebpRelativePath()),
+    $this->secondTestModel = TestModel::factory()->create([
+        'image' => $this->secondImage,
     ]);
+
+    Storage::disk()
+        ->assertExists($this->secondTestModel->image);
+
+    Storage::disk()
+        ->assertExists($this->secondImage);
 });
+
+it('can save an image by passing the path', function () {
+    $testImage = TestModel::find(1);
+
+    WebpService::make($testImage->image)->save();
+
+    Storage::disk()->assertExists($testImage->image);
+});
+
 
 it('must resize as needed', function () {
-    $testImage = TestModel::find(1);
+    $image = Image::make(Storage::get($this->testModel->image));
 
-    $path = $testImage->resize('image', 400, 400);
-
-    Storage::disk()
-        ->assertExists(ImageToWebp::toRelativePath($path));
-
-    Storage::disk()
-        ->assertExists(ImageToWebp::getOldImageRelativePath());
+    expect($image->width())->toBe(200)
+        ->and($image->height())->toBe(200);
 });
 
 it('must convert and overwrite all images in the directory ', function () {
+    $image = UploadedFile::fake()
+        ->image('image.jpg')
+        ->store('public');
+
+    $webpImage = Str::replace('.jpg', '.webp', $image);
+
     Artisan::call('public:to-webp --overwrite');
 
     Storage::disk()
-        ->assertExists($this->getTestImageWebpRelativePath());
-
-    // ensure that other files with other extensions are not deleted!
-    Storage::disk()
-        ->assertExists($this->getTempImageRelativePath());
-
-    // ensure that old files are deleted!
-    Storage::disk()
-        ->assertMissing($this->getTestImageRelativePath());
-});
-it('must convert and keep all images in the directory ', function () {
-    Artisan::call('public:to-webp');
-
-    Storage::disk()
-        ->assertExists($this->getTestImageWebpRelativePath());
-
-    // ensure that other files with other extensions are not deleted!
-    Storage::disk()
-        ->assertExists($this->getTempImageRelativePath());
+        ->assertExists($webpImage);
 
     // ensure that old files are there!
     Storage::disk()
-        ->assertExists($this->getTestImageRelativePath());
+        ->assertMissing($image);
+});
+it('must convert and keep all images in the directory ', function () {
+    $image = UploadedFile::fake()
+        ->image('image.jpg')
+        ->store('public');
+
+    $webpImage = Str::replace('.jpg', '.webp', $image);
+
+    Artisan::call('public:to-webp');
+
+    Storage::disk()->assertExists($webpImage);
+
+    // ensure that old files are there!
+    Storage::disk()
+        ->assertExists($image);
 });
 
 it('command must convert image field in the database ', function () {
-    Artisan::call('images:to-webp
-    EyadHamza\\\LaravelWebp\\\Tests\\\TestSupport\\\Models\\\TestModel
-    image');
+    $image = UploadedFile::fake()
+        ->image('image.jpg')
+        ->store('public');
 
+    $webpImage = Str::replace('.jpg', '.webp', $image);
+
+    DB::table('test_images')->insert([
+        'image' => $image,
+    ]);
+
+    Artisan::call('images:to-webp', [
+        'model' => TestModel::class,
+    ]);
 
     assertDatabaseHas('test_images', [
-        'image' => ImageToWebp::toFullPath($this->getTestImageWebpRelativePath()),
-    ]);
-
-    assertDatabaseMissing('test_images', [
-        'image' => ImageToWebp::toFullPath($this->getTestImageRelativePath()),
+        'image' => PathConversionSupport::appendWidthAndHeightToImageName($webpImage, 200, 200),
     ]);
 });
-test('command must convert all the image fields in the database ', function () {
-    Artisan::call('images:to-webp
-    EyadHamza\\\LaravelWebp\\\Tests\\\TestSupport\\\Models\\\TestModel');
 
-    assertDatabaseHas('test_images', [
-        'image' => ImageToWebp::toFullPath($this->getTestImageWebpRelativePath()),
-        'avatar' => ImageToWebp::toFullPath($this->getSecondTestImageWebpRelativePath()),
-    ]);
-
-    assertDatabaseMissing('test_images', [
-        'image' => ImageToWebp::toFullPath($this->getTestImageRelativePath()),
-    ]);
-});
-it('can support multiple image fields url in the database', function () {
-    $testImage = TestModel::find(1);
-
-    $testImage->saveImageAsWebp();
-
-    assertDatabaseHas('test_images', [
-        'image' => ImageToWebp::toFullPath($this->getTestImageWebpRelativePath()),
-        'avatar' => ImageToWebp::toFullPath($this->getSecondTestImageWebpRelativePath()),
-    ]);
-});
 
 it('should throw an exception if no image was given', function () {
-    $testImage = TestModel::create([
-        'image' => null,
-    ]);
     $this->expectException(NoImageGivenException::class);
 
-    ImageToWebp::make($testImage->image);
-
-    assertDatabaseHas('test_images', [
+    TestModel::create([
         'image' => null,
     ]);
+
 });
 it('should throw an exception if the path is for an not image was given', function () {
+    $this->expectException(NotImageException::class);
+
     $testImage = TestModel::create([
         'image' => 'hellothere',
     ]);
-    $this->expectException(NotImageException::class);
 
-    ImageToWebp::make($testImage->image);
-
-    assertDatabaseHas('test_images', [
-        'image' => null,
-    ]);
 });
-afterEach(fn () => $this->refreshAndClean());
+
+it('should not make the changes twice', function () {
+    $testImage = TestModel::find(1);
+
+    $oldPath = $testImage->image;
+
+    $testImage->update([
+        'image' => $testImage->image,
+    ]);
+
+    Storage::disk()->assertExists($oldPath);
+});
